refactor: fetching image in enum manifests
